requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
alembic
pydantic
passlib[bcrypt]
pyjwt
redis
rq
python-dotenv
requests
moviepy
python-multipart
